requirements.txt
opencv-python >= 4.5.3 
opencv-contrib-python >= 4.5.3
pandas >= 1.0.5
pillow >= 7.2.0
numpy >= 1.18.5